refactor: Update water image and description
The code changes in `water.json` and `menu.html` update the image and description for the "Water" drink option. The image is changed to `../images/water.jpg` and the description is updated to provide more detailed information about water. This improves the accuracy and clarity of the information presented to users.

Note: The commit message follows the conventional format of `<type>: <description>`.